[Bug 323006] [DB] Various PostgreSQL test failures If restarting, do not drop all tables (fixes durable locking tests)

+++ plugins/org.eclipse.emf.cdo.tests.db/src/org/eclipse/emf/cdo/tests/db/PostgresqlConfig.java
@@ -84,8 +84,10 @@ public class PostgresqlConfig extends DBConfig
     {
       OM.LOG.warn(ex.getMessage());
     }
-
-    dropDatabase();
+    if (!isRestarting())
+    {
+      dropDatabase();
+    }
 
     return dataSource;
   }
